fix: retry-after on the new 503, and correct two review comments
Closes the remaining PR #903 security-review threads:

- errors/openai.go: WriteAuthFailure now applies retry-after (and any
  other headers Authorize returns) on the upstream_unavailable branch,
  not just the rate-limit branch, via a small shared applyHeaders
  helper. Authorize now returns retry-after: 5 for that case. Without
  this, a 503 told the caller to retry in prose only, so SDK retry
  layers fell back to their own short backoff and hammered a
  control-plane that is by construction already unable to answer (LOW
  finding).

- authorizer.go: replaced the earlier, overconfident comment claiming
  the 401-vs-503 split adds no key-existence oracle. It does: control-
  plane's ResolveSnapshot fails fast for a nonexistent key (404, still
  401 here) but runs several more lookups for one that exists, any of
  which can fail under a degraded control-plane (503) -- so during
  exactly the kind of degraded window this PR is about, an existing key
  is more likely to 503 than a nonexistent one. Recorded as an accepted,
  narrow, low-value risk rather than a reason to revert (reverting
  restores the certain, severe cost this PR removes), with a suggested
  follow-up (rate-limit/aggregate repeated resolve failures per source)
  left as future work, not blocking here.

- client.go: corrected the comment on the 409 branch, which claimed
  behavior this endpoint does not have. handleInternalResolve only ever
  calls ResolveSnapshot, which never returns ErrRevoked/ErrDisabled/
  ErrNotActive (those three, and the 409 they produce, only happen on
  the account-facing key-management routes); a revoked/disabled/expired
  key resolves 200 here and is denied edge-side by CheckAccess. The
  branch itself stays as defensive classification, just no longer
  misdescribed.

- Test coverage the security review asked for: a WriteAuthFailure table
  case for upstream_unavailable -> 503 with retry-after, and one
  handler-level test per Authorize call site in inference/*.go
  (executeSync, executeStreaming, executeResponsesStreaming) proving
  each actually answers 503 rather than 401 -- closing the gap where
  every previous test stopped at the Client/Authorizer boundary and
  never exercised whether a call site actually used the shared mapper.
